+++ skin/src/main/res/drawable/rss_ic_circle_empty.xml
@@ -3,8 +3,8 @@
     android:shape="oval">
 
     <size
-        android:width="24dp"
-        android:height="24dp" />
+        android:width="@dimen/rss_activity_indicator_size"
+        android:height="@dimen/rss_activity_indicator_size" />
 
     <solid android:color="@android:color/transparent" />
